pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "meta-ads-cli"
version = "0.1.0"
description = "Production-grade CLI for Meta Ads management"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Meta Ads CLI" }]
dependencies = [
  "typer>=0.12.5",
  "rich>=13.9.4",
  "pydantic>=2.10.6",
  "PyYAML>=6.0.2",
  "facebook-business>=21.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-mock>=3.14.0",
  "ruff>=0.11.0",
]

[project.scripts]
meta-cli = "meta_cli.app:run"

[tool.hatch.build.targets.wheel]
packages = ["src/meta_cli"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
